Add files via upload
Dataset used is the Large Movie Review Dataset.
This is a dataset for binary sentiment classification containing substantially more data than previous benchmark datasets. We provide a set of 25,000 highly polar movie reviews for training and 25,000 for testing. There is additional unlabeled data for use as well. Raw text and an already processed bag of words formats are provided. See the README file contained in the release for more details.